add type

+++ unsupervised_learning/0x00-dimensionality_reduction/2-P_init.py
@@ -28,8 +28,8 @@ def P_init(X, perplexity):
 
     sum_x = np.sum(np.square(X), 1, keepdims=True)
     D = (np.add(np.add(-2 * np.dot(X, X.T), sum_x).T, sum_x))
-    P = np.zeros((n, n))
-    betas = np.ones((n, 1))
+    P = np.zeros((n, n), dtype="float64")
+    betas = np.ones((n, 1), dtype="float64")
     H = np.log2(perplexity)
 
     return (D, P, betas, H)
